another fix in odessa_test.cpp

+++ source/num/test/odessa_test.cpp
@@ -237,7 +237,7 @@ void run(void* data)
 {
   double* results = static_cast<double*>(data);
 
-  int nVar(3),nPar(3);
+  size_t nVar(3),nPar(3);
   
   double x[]={1,2,3,1,0,0,0,1,0,0,0,1};
   double p[]={8./3.,28,10};
